client implemented with a few methods
Implemented a client with a few working methods for starters:
* CreateTask	    <-- creates a task
* ListMyTasks	    <-- to support pagination
* ListAllMyTasks    <-- iterates for all your tasks

The rest are work in progress that I'll implement when I wake up
as it is 3:42PM and I started this work at around 1AM. I've got to
dig through their documentation to match up discrepancies as well as
study it.

For now, this suffices as a minimalistic API client.